requirements.txt
Flask==3.0.0
pandas==2.1.4
psycopg2-binary==2.9.9
pyarrow==14.0.2
requests==2.31.0
pyspark==3.5.0
boto3==1.34.13
localstack-utils==1.0.0
